requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
argcomplete==1.12.3
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.0.0
certifi==2020.6.20
cffi @ file:///C:/ci_310/cffi_1642682485096/work
chardet @ file:///D:/bld/chardet_1649184325588/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
clang==5.0
colorama==0.4.4
cryptography @ file:///C:/ci/cryptography_1639414690059/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
debugpy==1.5.0
decorator==5.1.0
entrypoints==0.3
flatbuffers==1.12
fonttools @ file:///D:/bld/fonttools_1649448633863/work
frozenlist==1.3.0
fsspec==2022.3.0
future @ file:///D:/bld/future_1649010380983/work
gast==0.4.0
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.1.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.8.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1633637554808/work
jsonpickle @ file:///home/conda/feedstock_root/build_artifacts/jsonpickle_1642174459935/work
jupyter-client==7.0.6
jupyter-core==4.8.1
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1644944417636/work
Markdown==3.3.4
matplotlib @ file:///D:/bld/matplotlib-suite_1639359207070/work
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
more-itertools @ file:///home/conda/feedstock_root/build_artifacts/more-itertools_1637732846337/work
multidict==6.0.2
munkres==1.1.4
music21 @ file:///home/conda/feedstock_root/build_artifacts/music21_1648773232336/work
nest-asyncio==1.5.1
numpy @ file:///C:/ci/numpy_and_numpy_base_1649782933444/work
oauthlib==3.2.0
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
parso==0.8.2
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.20
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyDeprecate==0.3.1
Pygments==2.10.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1649603503565/work
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.2
pytorch-lightning==1.4.5
pywin32==301
PyYAML @ file:///D:/bld/pyyaml_1648757357341/work
pyzmq==22.3.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
rsa==4.8
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.0
tensorflow-estimator==2.7.0
termcolor==1.1.0
torch==1.11.0
torchmetrics==0.8.0
torchvision==0.12.0
tornado==6.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1649051611147/work
traitlets==5.1.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
unicodedata2 @ file:///D:/bld/unicodedata2_1649112131705/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
wcwidth==0.2.5
webcolors @ file:///home/conda/feedstock_root/build_artifacts/webcolors_1606181549448/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1648847766614/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.12.1
yarl==1.7.2
zipp==3.6.0
